refactor: fully type combo box connector frontend file
The combo box connector was plain JavaScript, so nothing checked how it
uses the web component API. It is now TypeScript: the public API types
come from the @vaadin packages, and the private API it relies on is
declared in type-only files that ship with the connector. The connector
also moves to a component-named subfolder, as the conventions require.
