+++ vaadin-combo-box-flow-parent/vaadin-combo-box-flow-integration-tests/test/env-setup.ts
@@ -1,3 +1,3 @@
-window.Vaadin ||= {};
+window.Vaadin ||= {} as Vaadin;
 // @ts-expect-error
 window.Vaadin.Flow = {};
