+++ vaadin-combo-box-flow-parent/vaadin-combo-box-flow/src/main/java/com/vaadin/flow/component/combobox/ComboBox.java
@@ -92,7 +92,7 @@ import tools.jackson.databind.node.ObjectNode;
 @NpmPackage(value = "@vaadin/combo-box", version = "25.3.0-alpha8")
 @JsModule("@vaadin/combo-box/src/vaadin-combo-box.js")
 @JsModule("./flow-component-renderer.js")
-@JsModule("./comboBoxConnector.js")
+@JsModule("./vaadin-combo-box/comboBoxConnector.ts")
 public class ComboBox<T> extends ComboBoxBase<ComboBox<T>, T, T>
         implements HasPrefix, HasThemeVariant<ComboBoxVariant> {
