+++ vaadin-combo-box-flow-parent/vaadin-combo-box-flow/src/main/resources/META-INF/frontend/vaadin-combo-box/vaadin-combo-box-placeholder.d.ts
@@ -1,0 +1,5 @@
+// Declares the untyped @vaadin/combo-box placeholder module
+declare module '@vaadin/combo-box/src/vaadin-combo-box-placeholder.js' {
+  /** Placeholder object representing an item that is being loaded */
+  export class ComboBoxPlaceholder {}
+}
