+++ vaadin-combo-box-flow-parent/vaadin-combo-box-flow/tsconfig.json
@@ -1,0 +1,20 @@
+{
+  "compilerOptions": {
+    "target": "es2023",
+    "module": "esnext",
+    "moduleResolution": "bundler",
+    "strict": true,
+    "noImplicitReturns": true,
+    "noFallthroughCasesInSwitch": true,
+    "skipLibCheck": true,
+    "noEmit": true,
+    "allowImportingTsExtensions": true,
+    "useDefineForClassFields": false,
+    // Resolve @vaadin package types from the integration tests module, whose
+    // node_modules is populated by the Maven frontend build
+    "paths": {
+      "@vaadin/*": ["../vaadin-combo-box-flow-integration-tests/node_modules/@vaadin/*"]
+    }
+  },
+  "include": ["src/main/resources/META-INF/frontend/**/*"]
+}
